pyproject.toml
[project]
name = "hospital-management-system"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "flask>=3.1.2",
    "google-generativeai>=0.8.6",
    "ipykernel>=7.1.0",
    "ipython>=9.9.0",
    "langchain>=1.2.4",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.7",
    "langchain-google-genai>=4.2.0",
    "langchain-groq>=1.1.1",
    "langchain-huggingface>=1.2.0",
    "langchain-pinecone>=0.2.13",
    "langsmith>=0.6.4",
    "pymupdf>=1.26.7",
    "python-dotenv>=1.2.1",
    "qrcode[pil]>=8.2",
]
